+++ src/pages/login/index.tsx
@@ -1,139 +1,47 @@
-import React, { useState, FormEvent, ChangeEvent } from 'react';
-import type { JSX } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Link } from 'react-router-dom';
+import React, { useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
 import {
-  Box,
-  Container,
-  Typography,
-  Grid,
-  Stack,
-  Button,
   TextField,
-  Divider,
-  Paper,
+  Button,
+  Stack,
+  Box,
+  Typography,
   Alert,
+  Divider,
+  Grid,
 } from '@mui/material';
 
-export default function RegisterPage() {
+export default function LoginPage() {
   const navigate = useNavigate();
   
   type FormErrors = {
-    name?: string;
-    phone?: string;
     email?: string;
     password?: string;
-    confirmPassword?: string;
-    captcha?: string;
     general?: string;
   };
 
-  const [registerMode, setRegisterMode] = useState<'email' | 'phone'>('email');
   const [formData, setFormData] = useState({
-    name: '',
     email: '',
-    phone: '',
     password: '',
-    confirmPassword: '',
-    company: '',
-    captcha: ''
   });
   const [errors, setErrors] = useState<FormErrors>({});
   const [loading, setLoading] = useState(false);
-  const [captchaLoading, setCaptchaLoading] = useState(false);
 
   // 表单验证逻辑
   const validate = () => {
     let validationErrors: FormErrors = {};
     
-    // 用户名验证
-    if (!formData.name.trim()) {
-      validationErrors.name = '请输入用户名';
+    // 邮箱校验逻辑
+    if (!/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(formData.email)) {
+      validationErrors.email = '请输入正确邮箱格式';
     }
     
-    // 手机号/邮箱校验逻辑
-    if (registerMode === 'phone') {
-      if (!/^1[3-9]\d{9}$/.test(formData.phone)) {
-        validationErrors.phone = '请输入正确手机号';
-      }
-    } else {
-      // 更严格的邮箱格式验证
-      if (!/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(formData.email)) {
-        validationErrors.email = '请输入正确邮箱格式';
-      }
-    }
-    
-    // 密码强度校验
-    if (formData.password.length < 8) {
-      validationErrors.password = '密码需至少8位';
-    } else if (!/[A-Z]/.test(formData.password)) {
-      validationErrors.password = '需包含大写字母';
-    }
-    
-    // 确认密码校验
-    if (formData.password !== formData.confirmPassword) {
-      validationErrors.confirmPassword = '两次输入的密码不一致';
-    }
-    
-    // 验证码校验（如果有验证码字段）
-    if (formData.captcha && !/^\d{6}$/.test(formData.captcha)) {
-      validationErrors.captcha = '请输入6位数字验证码';
+    // 密码校验
+    if (!formData.password) {
+      validationErrors.password = '请输入密码';
     }
     
     return validationErrors;
-  };
-
-  // 发送验证码
-  const sendCaptcha = async () => {
-    // 验证手机号或邮箱
-    let currentErrors = {...errors};
-    let hasError = false;
-    
-    if (registerMode === 'phone') {
-      if (!/^1[3-9]\d{9}$/.test(formData.phone)) {
-        currentErrors.phone = '请输入正确手机号';
-        hasError = true;
-      } else {
-        delete currentErrors.phone;
-      }
-    } else {
-      if (!/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(formData.email)) {
-        currentErrors.email = '请输入正确邮箱格式';
-        hasError = true;
-      } else {
-        delete currentErrors.email;
-      }
-    }
-    
-    if (hasError) {
-      setErrors(currentErrors);
-      return;
-    }
-
-    setCaptchaLoading(true);
-    try {
-      // 发送验证码请求
-      const response = await fetch('/api/send-captcha', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          mode: registerMode,
-          [registerMode]: registerMode === 'phone' ? formData.phone : formData.email
-        }),
-      });
-
-      if (!response.ok) {
-        throw new Error('验证码发送失败');
-      }
-      
-      alert('验证码已发送');
-    } catch (error) {
-      alert('验证码发送失败，请重试');
-    } finally {
-      setCaptchaLoading(false);
-    }
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
@@ -149,7 +57,7 @@ export default function RegisterPage() {
     setLoading(true);
 
     try {
-      const response = await fetch('/api/auth/register', {
+      const response = await fetch('/api/auth/login', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
@@ -160,14 +68,14 @@ export default function RegisterPage() {
       const data = await response.json();
 
       if (!response.ok) {
-        throw new Error(data.message || '注册失败');
+        throw new Error(data.message || '登录失败');
       }
 
-      navigate('/login?registered=true');
+      navigate('/dashboard');
     } catch (err) {
       setErrors({
         ...errors,
-        general: err instanceof Error ? err.message : '注册失败'
+        general: err instanceof Error ? (err.message.includes('密码') ? '密码错误' : err.message) : '密码错误'
       });
     } finally {
       setLoading(false);
@@ -188,84 +96,32 @@ export default function RegisterPage() {
             />
           </div>
           <h2 className="text-3xl font-bold text-gray-900 dark:text-white">
-            创建新账号
+            登录账号
           </h2>
           <p className="mt-2 text-sm text-gray-600 dark:text-gray-400">
-            已有账号？{' '}
-            <Link to="/login" className="font-medium text-blue-600 hover:text-blue-500 dark:text-blue-400">
-              立即登录
+            还没有账号？{' '}
+            <Link to="/register" className="font-medium text-blue-600 hover:text-blue-500 dark:text-blue-400">
+              立即注册
             </Link>
           </p>
         </div>
 
-        {/* 注册模式切换 */}
-        <div className="flex rounded-md shadow-sm mt-4">
-          <button
-            type="button"
-            onClick={() => setRegisterMode('email')}
-            className={`flex-1 py-2 px-4 text-sm font-medium rounded-l-lg border ${registerMode === 'email' 
-              ? 'bg-blue-50 border-blue-500 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300' 
-              : 'bg-white border-gray-300 text-gray-700 dark:bg-gray-800 dark:border-gray-600 dark:text-gray-300'}`}
-          >
-            邮箱注册
-          </button>
-          <button
-            type="button"
-            onClick={() => setRegisterMode('phone')}
-            className={`flex-1 py-2 px-4 text-sm font-medium rounded-r-lg border ${registerMode === 'phone' 
-              ? 'bg-blue-50 border-blue-500 text-blue-700 dark:bg-blue-900/30 dark:text-blue-300' 
-              : 'bg-white border-gray-300 text-gray-700 dark:bg-gray-800 dark:border-gray-600 dark:text-gray-300'}`}
-          >
-            手机号注册
-          </button>
-        </div>
-
         <form className="mt-6 space-y-6 flex flex-col items-center" onSubmit={handleSubmit}>
           <Stack spacing={3} sx={{ width: '30%' }}>
-            {/* 用户名输入框 */}
+            {/* 邮箱输入框 */}
             <TextField
               required
               fullWidth
-              id="name"
-              name="name"
-              label="用户名"
-              value={formData.name}
-              onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-              error={!!errors.name}
-              helperText={errors.name}
+              id="email"
+              name="email"
+              label="邮箱地址"
+              type="email"
+              value={formData.email}
+              onChange={(e) => setFormData({ ...formData, email: e.target.value })}
+              error={!!errors.email}
+              helperText={errors.email}
               variant="outlined"
             />
-
-            {/* 邮箱/手机号输入框 */}
-            {registerMode === 'email' ? (
-              <TextField
-                required
-                fullWidth
-                id="email"
-                name="email"
-                label="邮箱地址"
-                type="email"
-                value={formData.email}
-                onChange={(e) => setFormData({ ...formData, email: e.target.value })}
-                error={!!errors.email}
-                helperText={errors.email}
-                variant="outlined"
-              />
-            ) : (
-              <TextField
-                required
-                fullWidth
-                id="phone"
-                name="phone"
-                label="手机号码"
-                type="tel"
-                value={formData.phone}
-                onChange={(e) => setFormData({ ...formData, phone: e.target.value })}
-                error={!!errors.phone}
-                helperText={errors.phone}
-                variant="outlined"
-              />
-            )}
 
             {/* 密码输入框 */}
             <TextField
@@ -278,59 +134,18 @@ export default function RegisterPage() {
               value={formData.password}
               onChange={(e) => setFormData({ ...formData, password: e.target.value })}
               error={!!errors.password}
-              helperText={errors.password || '至少8位，包含大写字母'}
+              helperText={errors.password}
               variant="outlined"
             />
 
-            {/* 确认密码输入框 */}
-            <TextField
-              required
-              fullWidth
-              id="confirmPassword"
-              name="confirmPassword"
-              label="确认密码"
-              type="password"
-              value={formData.confirmPassword}
-              onChange={(e) => setFormData({ ...formData, confirmPassword: e.target.value })}
-              error={!!errors.confirmPassword}
-              helperText={errors.confirmPassword}
-              variant="outlined"
-            />
-
-            {/* 验证码输入框 */}
-            <Box sx={{ display: 'flex', gap: 1 }}>
-              <TextField
-                sx={{ flexGrow: 1 }}
-                id="captcha"
-                name="captcha"
-                label="验证码"
-                value={formData.captcha}
-                onChange={(e) => setFormData({ ...formData, captcha: e.target.value })}
-                error={!!errors.captcha}
-                helperText={errors.captcha}
-                variant="outlined"
-                inputProps={{ maxLength: 6 }}
-              />
-              <Button
-                variant="contained"
-                onClick={sendCaptcha}
-                disabled={captchaLoading}
-                sx={{ minWidth: 120, height: 56 }}
+            <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
+              <Link
+                to="/forgot-password"
+                className="text-sm text-blue-600 hover:text-blue-500 dark:text-blue-400"
               >
-                {captchaLoading ? '发送中...' : '获取验证码'}
-              </Button>
+                忘记密码？
+              </Link>
             </Box>
-
-            {/* 公司名称输入框 */}
-            <TextField
-              fullWidth
-              id="company"
-              name="company"
-              label="公司名称（选填）"
-              value={formData.company}
-              onChange={(e) => setFormData({ ...formData, company: e.target.value })}
-              variant="outlined"
-            />
           </Stack>
 
           {errors.general && (
@@ -377,25 +192,7 @@ export default function RegisterPage() {
                   }}
                 />
               ) : (
-                <>
-                  注册
-                  <Box
-                    component="svg"
-                    sx={{
-                      width: 14,
-                      height: 14,
-                      transition: 'transform 0.2s',
-                      '.group:hover &': {
-                        transform: 'translateX(4px)'
-                      }
-                    }}
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14 5l7 7m0 0l-7 7m7-7H3" />
-                  </Box>
-                </>
+                '登录'
               )}
             </Button>
           </Box>
@@ -404,7 +201,7 @@ export default function RegisterPage() {
         <Box sx={{ width: '30%', mt: 6, position: 'relative' }}>
           <Divider>
             <Typography variant="body2" color="text.secondary">
-              其他注册方式
+              其他登录方式
             </Typography>
           </Divider>
 
@@ -421,7 +218,7 @@ export default function RegisterPage() {
                   }
                 }}
               >
-                微信注册
+                微信登录
               </Button>
             </Grid>
             <Grid item xs={6}>
@@ -436,7 +233,7 @@ export default function RegisterPage() {
                   }
                 }}
               >
-                QQ注册
+                QQ登录
               </Button>
             </Grid>
           </Grid>
